remove go1.7 support
