+++ .travis.yml
@@ -1,7 +1,6 @@
 language: go
 sudo: false
 go:
-  - 1.7.x
   - 1.8.x
   - 1.9.x
   - 1.10.x
