+++ examples/graceful-shutdown/close/server.go
@@ -1,5 +1,3 @@
-// +build go1.8
-
 package main
 
 import (
