+++ examples/graceful-shutdown/graceful-shutdown/server.go
@@ -1,5 +1,3 @@
-// +build go1.8
-
 package main
 
 import (
