--- response_writer_1.7.go
@@ -1,12 +1,0 @@
-// Copyright 2018 Gin Core Team.  All rights reserved.
-// Use of this source code is governed by a MIT style
-// license that can be found in the LICENSE file.
-
-// +build !go1.8
-
-package gin
-
-// ResponseWriter ...
-type ResponseWriter interface {
-	responseWriterBase
-}
